convert the task manager into a scheduler
