+++ pkg/tenant/topics.go
@@ -231,7 +231,7 @@ func (s *Server) ProjectTopicCreate(c *gin.Context) {
 	}
 
 	// Update project stats in the background
-	s.tasks.QueueContext(middleware.TaskContext(c), radish.Func(func(ctx context.Context) error {
+	s.tasks.QueueContext(middleware.TaskContext(c), radish.TaskFunc(func(ctx context.Context) error {
 		return s.UpdateProjectStats(ctx, userID, t.ProjectID)
 	}), radish.WithErrorf("could not update stats for project %s", t.ProjectID.String()))
